// BOnlineStore.UI/ClientApp/src/app/enums/production/formula-variable-type.enum.ts
export enum FormulaVariableTypeEnums {
  WIDTH1 = 'WIDTH1',
  WIDTH2 = 'WIDTH2',
  WIDTH3 = 'WIDTH3',
  HEIGHT = 'HEIGHT',
  CONSTANT = 'CONSTANT',
  RESULTVARIABLE = 'RESULTVARIABLE',
  OPENPARENTHESIS = '(',
  CLOSEPARENTHESIS = ')',
  PLUS = '+',
  MINUS = '-',
  MULTIPLY = '*',
  DIVIDE = '/',
}